+++ website/docker-entrypoint.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+set -eu
+
+if [ -z "${PORT:-}" ]; then
+  export PORT=8080
+fi